feat(text): delete char assembler
