+++ packages/core/src/2d/atlas/FontAtlas.ts
@@ -1,7 +1,7 @@
 import { RefObject } from "../../asset/RefObject";
 import { Engine } from "../../Engine";
 import { Texture2D } from "../../texture/Texture2D";
-import { CharInfo } from "../assembler/CharInfo";
+import { CharInfo } from "../text/CharInfo";
 
 /**
  * @internal
