+++ packages/core/src/2d/text/CharInfo.ts
