+++ packages/core/src/2d/text/CharRenderData.ts
