+++ packages/core/src/2d/text/CharRenderDataPool.ts
